+++ src/components/probe/prober/mariadb/request.ts
@@ -94,3 +94,7 @@ async function checkConnection(params?: MariaParam) {
   await client.end()
   return true
 }
+
+export const moduleExports = {
+  mariaRequest,
+}
