+++ test/index.test.ts
@@ -26,7 +26,7 @@ import path from 'node:path'
 import { expect, test } from '@oclif/test'
 import { use } from 'chai'
 import spies from 'chai-spies'
-import { stub, type SinonStub } from 'sinon'
+import sinon from 'sinon'
 import cmd from '../src/commands/monika.js'
 import * as IpUtil from '../src/utils/public-ip.js'
 
@@ -35,10 +35,10 @@ const { resolve } = path
 use(spies)
 
 describe('monika', () => {
-  let getPublicIPStub: SinonStub
+  let getPublicIPStub: sinon.SinonStub
 
   beforeEach(() => {
-    getPublicIPStub = stub(IpUtil, 'getPublicIp')
+    getPublicIPStub = sinon.stub(IpUtil, 'getPublicIp')
   })
 
   afterEach(() => {
